Moved functions for running contract tests into the ScenarioTest class
