+++ core/src/test/kotlin/in/specmatic/core/ScenarioAsTest.kt
@@ -14,7 +14,7 @@ import org.junit.jupiter.api.Test
 import java.util.*
 import java.util.function.Consumer
 
-internal class ScenarioTest {
+internal class ScenarioAsTest {
     @Test
     fun `should generate one test scenario when there are no examples`() {
         val scenario = Scenario(
@@ -516,14 +516,14 @@ paths:
         val contractTestScenarios = contract.generateContractTests(emptyList())
 
         val result: Result =
-            executeTestAndReturnResultAndResponse(contractTestScenarios.first(), object : TestExecutor {
-                override fun execute(request: HttpRequest): HttpResponse {
-                    return HttpResponse.ok("abc")
-                }
+            contractTestScenarios.first().runTest(object : TestExecutor {
+                        override fun execute(request: HttpRequest): HttpResponse {
+                            return HttpResponse.ok("abc")
+                        }
 
-                override fun setServerState(serverState: Map<String, Value>) {
-                }
-            }, DefaultStrategies).first as Result.Failure
+                        override fun setServerState(serverState: Map<String, Value>) {
+                        }
+                    }).first as Result.Failure
 
         assertThat(result.reportString()).contains("Contract expected")
         assertThat(result.reportString()).contains("response contained")
